setup-cli/requirements.txt
rich>=13.0.0
click>=8.1.0
paramiko>=3.4.0
pyyaml>=6.0.0
